requirements.txt
telethon==1.40.0
cryptography==45.0.4
httpx==0.28.1
yt-dlp==2026.03.31
Pillow==11.2.1
qrcode==8.2
